cleaned some stuff up. SSH server starts on app start... pressing seefood sends filename to SSH for verification. ----------------------------------------------------------------------- I created a new class SSH_SERVER_Testing. This class runs the tensorflow code and allows for very fast image processing via the addFile command
SHIT WORKS STILL NEEDS TO BE PRETTY BUT IT WORKS!!! THANKS DAN

NOTES:
1.In Android Studio watch the logcat Verbos setting and search for MyApp

------------------------------------------------------------------------
Press SeeFood to submit the currently hardcoded picture to the server
and you can see the ssh terminal communicate back via the logcat. The
last line of the ssh message is saved into the string lastLine inside
the server class. If you press the camera button I have it currently
configured to TOAST the last line(Which should always be the the pair
results from TensorFlow)